mcp/biomcp-ollama-chat/webapp/requirements.txt
flask==3.0.0
httpx==0.27.0
mcp==1.0.0
asyncio==3.4.3
python-dotenv==1.0.0
aiohttp==3.9.1